requirements.txt
astroid==2.0.1
Django==2.0.7
djangorestframework==3.8.2
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==2.0.1
pylint-django==0.11.1
pylint-plugin-utils==0.4
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
